Add an escape-hatch yaml file that maps packages to SPDX licenses
Some components cannot be parsed by this license checker and this
creates a churn when upgrading component versions, making such upgrades
dependent on new versions of this checker. Typical scenario

1. We want to release a new version of component x.
2. The new release must use the latest of component y.
3. The go-mkopensource refuses to parse something that the y points to.
4. Our release is blocked on a new mk-opensource release.

This addition removes the block by making it possible for component x
to specify the SPDX licenses that component y is using.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>