go: add ignore dirty flag
Currently, the go-mkopensource will err out and not generate an output
if it detects that the go.mod/go.sum are dirty. This is the ideal flow
for CI but during the normal dev loop it is common to make changes
to the go.mod and then run a "generate" command.

This flag allows the dirty check to be ignored when added so that it
will generate it normally.

Signed-off-by: Lance Austin <laustin@datawire.io>